Remove classes não utilizadas
